fix(migration): change createdAt, updatedAt, and deletedAt fields to bigInteger type

+++ database/migrations/2026_05_21_190449_create_contacts_table.php
@@ -56,9 +56,9 @@ return new class extends Migration
 
             $table->string('code')->nullable();
 
-            $table->integer('createdAt')->nullable();
-            $table->integer('updatedAt')->nullable();
-            $table->integer('deletedAt')->nullable();
+            $table->bigInteger('createdAt')->nullable();
+            $table->bigInteger('updatedAt')->nullable();
+            $table->bigInteger('deletedAt')->nullable();
 
             /*
             |--------------------------------------------------------------------------
